platform/requirements.txt
# Macaron Agent Platform - Dependencies
fastapi>=0.104
uvicorn[standard]>=0.24
jinja2>=3.1
python-multipart>=0.0.6
aiohttp>=3.8
aiohttp-sse>=2.0
pydantic>=2.0
pyyaml>=6.0
watchdog>=3.0
httpx>=0.27
markdown>=3.5
psycopg[binary]>=3.1
psycopg-pool>=3.1
psutil>=5.9
numpy>=1.26
bcrypt>=4.0
PyJWT>=2.8
slowapi>=0.1.9
# OpenTelemetry (opt-in via OTEL_ENABLED=1)
opentelemetry-api>=1.20
opentelemetry-sdk>=1.20
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-exporter-otlp-proto-http>=1.20

# File upload parsing
pypdf>=4.0
python-docx>=1.1

# Browser Web Push (optional — for browser push notifications)
pywebpush>=2.0

# Redis pub/sub — cross-process event bus (IHM ↔ Factory decoupling)
redis[asyncio]>=5.0

# LSP code intelligence — symbol navigation and diagnostics (lsp_tools)
jedi>=0.19
